# web/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "BFR WEBUI GO API",
    "description": "Ultra-lightweight Android System Control Panel & WebUI REST API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/auth/login": {
      "post": {
        "summary": "Authenticate session",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "password": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Login successful" },
          "401": { "description": "Invalid password" }
        }
      }
    },
...
    "/api/power": {
      "post": {
        "summary": "Execute system power action",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "action": { "type": "string", "enum": ["reboot", "shutdown", "soft_reboot", "recovery", "bootloader"] }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Power action initiated" },
          "400": { "description": "Invalid action" }
        }
      }
    },
...
    "/api/files/list": {
      "get": {
        "summary": "List files in directory",
        "parameters": [
          { "name": "path", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "File list" } }
      }
    },
...
  }
}